pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["md <marceldiomare@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.15"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.27.0"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
redis = "^5.1.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
d = "^0.2.2"
django-environ = "^0.11.2"
django-redis = "^5.4.0"
flake8 = "^7.1.1"
setuptools = "^75.1.0"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"

[tool.poetry.group.lint.dependencies]
black = "^23.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"